requirements.txt
streamlit
psycopg2-binary
pandas
plotly
openpyxl
fpdf
matplotlib